General: Various fixes to the correctness of code and documentation reported by PHPStan.
Props justlevine

See #63268
Built from https://develop.svn.wordpress.org/trunk@60275


git-svn-id: http://core.svn.wordpress.org/trunk@59611 1a063a9b-81f0-0310-95a4-ce76da25c4cd

+++ wp-includes/category-template.php
@@ -793,7 +793,7 @@ function wp_tag_cloud( $args = '' ) {
  * @return int Scaled count.
  */
 function default_topic_count_scale( $count ) {
-	return round( log10( $count + 1 ) * 100 );
+	return (int) round( log10( $count + 1 ) * 100 );
 }
 
 /**
